pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "pawabot"
version = "0.1.0"
description = "A bot for many things: aria2 management, torrent sites crawling, media organization with filebot and plex."
authors = ["Timothée Mazzucotelli <pawamoy@pm.me>"]
license = "ISC License"
readme = "README.md"
repository = "https://github.com/pawamoy/pawabot"
homepage = "https://github.com/pawamoy/pawabot"
keywords = ["telegram-bot", "aria2"]
packages = [ { include = "pawabot", from = "src" } ]
include = [
    "README.md",
    "pyproject.toml"
]

[tool.poetry.dependencies]
python = "^3.6"
python-telegram-bot = {version = "=12.0.0b1",allows-prereleases = true}
sqlalchemy = "^1.3"
aria2p = "^0.6.0"
beautifulsoup4 = "^4.8"
requests = "^2.22"
tomlkit = "^0.5.5"
privibot = "^0.1.0"
loguru = "^0.3.2"
xdg = "^4.0"

[tool.poetry.dev-dependencies]
bandit = "^1.5"
black = { version = "*", allows-prereleases = true }
coverage = { version = "=5.0a8", allows-prereleases = true }
flake8 = "^3.6"
ipython = "^7.2"
isort = { version = "^4.3", extras = ["pyproject"] }
jinja2-cli = "^0.7.0"
pylint = { git = "https://github.com/PyCQA/pylint" }
pytest = "^4.3"
pytest-cov = "^2.8"
pytest-sugar = "^0.9.2"
pytest-xdist = "^1.26"
recommonmark = "^0.6.0"
safety = "^1.8"
sphinx = "^1.8"
sphinx-rtd-theme = "^0.4.2"
toml = "^0.10.0"
cruft = "^1.1"

[tool.poetry.scripts]
pawabot = "pawabot.cli:main"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
not_skip = "__init__.py"
multi_line_output = 3
force_single_line = false
balanced_wrapping = true
default_section = "THIRDPARTY"
known_first_party = "pawabot"
include_trailing_comma = true

[tool.pylint.options]
disable = """
  bad-continuation,
  too-many-lines,
  too-many-public-methods,
  too-many-locals,
  too-many-statements,
  too-many-arguments,
  too-many-instance-attributes,
  too-many-branches,
  too-few-public-methods,
  duplicate-code
"""
max-line-length = "120"
